+++ Informatik/Informatik/graphics/Window.cpp
@@ -81,7 +81,7 @@ Window::Window() // Load from file, or if not found w = 50 & h = 50
     level->addEntity(slimeo);
     level->addEntity(slimer);
     
-    Projectile *projectile = new Projectile(0, 0, PI * 15.0 / 8.0);
+    Projectile *projectile = new Projectile(0, 0, (float) PI * 15.0f / 8.0f);
     level->addEntity(projectile);
     
     
@@ -200,7 +200,7 @@ void Window::runGameLoop()
                 {
                     for(int i = 0; i < 5; i++) // Shoot n projectiles
                     {
-                        ExplodingProjectile *p = new ExplodingProjectile(NORMAL, level->player->x_pos, level->player->y_pos, TO_RAD(rand() % 360));
+                        ExplodingProjectile *p = new ExplodingProjectile(NORMAL, level->player->x_pos, level->player->y_pos, (float) TO_RAD(rand() % 360));
                         level->addEntity(p);
                     }
                 }
@@ -212,7 +212,7 @@ void Window::runGameLoop()
                 {
                     int xdif = (int) (e.button.x / SCALE_X) - PLAYER_OFFSET_X + level->player->xoff;
                     int ydif = (int) (e.button.y / SCALE_Y) - PLAYER_OFFSET_Y + level->player->yoff;
-                    float angle = -atan2(ydif, xdif);
+                    float angle = -(float) atan2(ydif, xdif);
                     
                     Projectile *p = new Projectile(level->player->x_pos, level->player->y_pos, angle);
                     level->addEntity(p);
