+++ Informatik/Informatik/util/SDL_Util.hpp
@@ -41,7 +41,7 @@
 #define COLOR(r, col) SDL_SetRenderDrawColor(r, ((col) >> 16) & 0xFF, ((col) >> 8) & 0xFF, (col) & 0xFF, ((col) >> 24) & 0xFF)
 #define TO_COLOR(col) SDL_Color({ (uint8_t) ((col) >> 16), (uint8_t) ((col) >> 8), (uint8_t) (col), (uint8_t) ((col) >> 24) })
 
-#define TRANSFORM_LEVEL_POS(rect, xoff, yoff) {rect.x += xoff + PLAYER_OFFSET_X; rect.y += yoff + PLAYER_OFFSET_Y;}
+#define TRANSFORM_LEVEL_POS(rect, xoff, yoff) {rect.x += (int) xoff + PLAYER_OFFSET_X; rect.y += (int) yoff + PLAYER_OFFSET_Y;}
 
 extern TTF_Font *font;
 extern float SCALE_X, SCALE_Y;
